同步

+++ scripts/pack.js
@@ -20,11 +20,13 @@ optParser.addOption('e', 'encrypt', 'boolean', 'Whether encrypt during pack (Eg.
 optParser.addOption('d', 'debug', 'boolean', '(Disabled)');
 optParser.addOption('o', 'addon-debug', 'boolean', 'Whether pack debug addon (Eg. pack.js -t webrtc-agent -o)');
 optParser.addOption('f', 'full', 'boolean', 'Whether perform a full pack (--full is the equalivation of pack.js -t all -r -i)');
-optParser.addOption('s', 'sample-path', 'string', 'Specify sample path (Eg. pack.js -t all -s ${samplePath})');
+optParser.addOption('p', 'app-path', 'string', 'Specify app path (Eg. pack.js -t all --app-path ${appPath})');
 optParser.addOption('a', 'archive', 'string', 'Specify archive name (Eg. pack.js -t all -a ${archiveName})');
 optParser.addOption('n', 'node-module-path', 'string', 'Specify shared-node-module directory');
 optParser.addOption('c', 'copy-module-path', 'string', 'Specify copy node modules directory');
 optParser.addOption('b', 'binary', 'boolean', 'Pack binary');
+optParser.addOption('np', 'no-pseudo', 'boolean', 'Whether use pseudo library');
+optParser.addOption('wf', 'with-ffmpeg', 'boolean', 'Whether pack ffmpeg library');
 optParser.addOption('h', 'help', 'boolean', 'Show help');
 
 const options = optParser.parseArgs(process.argv);
@@ -71,14 +73,15 @@ if (options.encrypt) {
 
   // Check encrypt deps
   console.log('Checking encrypt dependencies...');
-  try {
-    for (const dep of encryptDeps) {
-      execSync(`npm list -g ${dep}`);
-    }
+  const npmRoot = execSync(`npm root -g`).toString().trim();
+  const missingDeps = encryptDeps.filter((dep) => {
+    return !fs.existsSync(path.join(npmRoot, dep));
+  });
+
+  if (missingDeps.length === 0) {
     console.log('Encrypt dependencies OK.');
-  } catch (e) {
-    console.log('Install node dependencies for Encrypt...');
-    for (const dep of encryptDeps) {
+  } else {
+    for (const dep of missingDeps) {
       execSync(`npm install -g --save-dev ${dep}`, { stdio: 'inherit' });
     }
   }
@@ -323,25 +326,31 @@ function packAddon(target) {
     .then((libsOk) => {
       // Replace openh264 if needed
       let libOpenh264 = path.join(libDist, 'libopenh264.so.4');
-      if (options['archive'] && fs.existsSync(libOpenh264)) {
+      if (options['archive'] && fs.existsSync(libOpenh264) && !options['no-pseudo']) {
         let dummyOpenh264 = path.join(rootDir, 'third_party/openh264/pseudo-openh264.so');
         execSync(`cp -av ${dummyOpenh264} ${libOpenh264}`);
       }
-      if (target.rules.name.indexOf('video') === 0) {
-        let vasrc = path.join(depsDir, 'bin/vainfo');
-        let vadist = path.join(packDist, 'bin');
-        if (fs.existsSync(vasrc)) {
-          if (!fs.existsSync(vadist)) {
-            execSync(`mkdir -p ${vadist}`);
-          }
-          execSync(`cp -av ${vasrc} ${vadist}`);
-        }
-      }
-      if (osType.includes('ubuntu')) {
-        let libSvtHevcEnc = path.join(libDist, 'libSvtHevcEnc.so.1');
+
+      let libSvtHevcEnc = path.join(libDist, 'libSvtHevcEnc.so.1');
+      if (options['archive'] && fs.existsSync(libSvtHevcEnc) && !options['no-pseudo']) {
         let dummySvtHevcEnc = path.join(rootDir, 'third_party/SVT-HEVC/pseudo-svtHevcEnc.so');
         execSync(`cp -av ${dummySvtHevcEnc} ${libSvtHevcEnc}`);
       }
+
+      if (target.rules.name.indexOf('video') === 0
+        || target.rules.name.indexOf('analytics') === 0) {
+        if (fs.existsSync('/opt/intel/mediasdk')) {
+          let dst_bin = path.join(packDist, 'bin');
+          let dst_lib = path.join(packDist, 'lib');
+          if (!fs.existsSync(dst_bin)) {
+            execSync(`mkdir -p ${dst_bin}`);
+          }
+
+          execSync(`find /opt/intel/mediasdk -type f -name metrics_monitor  | xargs -I '{}' cp -av '{}' ${dst_bin}`);
+          execSync(`find /opt/intel/mediasdk -type f -name libcttmetrics.so | xargs -I '{}' cp -av '{}' ${dst_lib}`);
+        }
+      }
+
       console.log(target.rules.name, '- Pack addon finished.');
     });
 }
@@ -357,13 +366,8 @@ function getAddonLibs(addonPath) {
   env['LD_LIBRARY_PATH'] = (env['LD_LIBRARY_PATH'] || '');
   env['LD_LIBRARY_PATH'] = path.join(depsDir, 'lib') +
     ':' + path.join(rootDir, 'third_party/openh264') +
-    ':' + path.join(rootDir, 'third_party/re') +
+    ':' + path.join(rootDir, 'third_party/quic-lib/dist/lib') +
     ':' + env['LD_LIBRARY_PATH'];
-
-    if (osType.includes('ubuntu')) {
-        env['LD_LIBRARY_PATH'] = ':' + path.join(rootDir, 'third_party/SVT-HEVC') +
-            ':' + env['LD_LIBRARY_PATH'];
-    }
 
   return exec(`ldd ${addonPath} | grep '=>' | awk '{print $3}' | sort | uniq | grep -v "^("`, { env })
     .then((stdout, stderr) => {
@@ -389,6 +393,17 @@ function getAddonLibs(addonPath) {
                   return line;
                 }).catch((e) => line);
             }
+          })
+          .catch((e) => {
+            // give more detail when ldd returns somelib.so => not found
+            if (!line.startsWith('/')) {
+              return exec(`ldd ${addonPath} | grep '=>' | grep -v '=> /'`).then(stdout => {
+                e.message = `library dependency not found for\n  ${addonPath}:\n${stdout}` +
+                  'Something failed to build. Try nvm use v8.15.0 and rerun build.js.';
+                throw e;
+              });
+            }
+            throw e;
           });
         checks.push(checkPros[line]);
       }
@@ -404,15 +419,20 @@ function isLibAllowed(libSrc) {
     return false;
 
   const whiteList = [
+    'rtcadapter',
     'libnice',
     'libSvtHevcEnc',
     'libusrsctp',
-    'libav',
-    'libsw',
     'libopenh264',
     'libre',
     'sipLib',
+    'librawquic'
   ];
+  if (!options['archive'] || options['with-ffmpeg']) {
+    whiteList.push('libav');
+    whiteList.push('libsw');
+  }
+
   const libName = path.basename(libSrc);
 
   var found = false;
@@ -588,24 +608,37 @@ function packScripts() {
   execSync(`chmod +x ${binDir}/\*.sh`);
 }
 
-function packSamples() {
-  if (!options['sample-path']) return;
+function packApps() {
+  if (!options['app-path']) return;
   chdir(originCwd);
-  var samplePath = options['sample-path'];
-  if (!fs.existsSync(samplePath)) {
-    console.log(`\x1b[31mError: ${samplePath} does not exist\x1b[0m`);
+  var appPath = options['app-path'];
+  if (!fs.existsSync(appPath)) {
+    console.log(`\x1b[31mError: ${appPath} does not exist\x1b[0m`);
     return;
   }
-  execSync(`rm -rf ${distDir}/extras`);
-  execSync(`mkdir -p ${distDir}/extras`);
-  execSync(`cp -a ${samplePath} ${distDir}/extras/basic_example`);
+  execSync(`rm -rf ${distDir}/apps`);
+  execSync(`mkdir -p ${distDir}/apps`);
+  console.log('\x1b[32mApps folder created in :', distDir, '\x1b[0m');
+  execSync(`cp -a ${appPath} ${distDir}/apps/current_app`);
 
-  const certScript = `${distDir}/extras/basic_example/initcert.js`;
+  // Look in the app's package.json to see what file to use for main.js
+  var jsonTXT = execSync(`cat ${distDir}/apps/current_app/package.json`);
+  var appJSON = JSON.parse(jsonTXT)["main"];
+
+  if (!appJSON === undefined) {
+    console.log("\x1b[31mError: No main js file for the app\x1b[0m");
+    return;
+  } else {
+    // Make a soft link to the main JS file node.js should call
+    if (appJSON !== 'main.js')
+      execSync(`ln -srf ${distDir}/apps/current_app/${appJSON} ${distDir}/apps/current_app/main.js`);
+  }
+  const certScript = `${distDir}/apps/current_app/initcert.js`;
   if (fs.existsSync(certScript))
     execSync(`chmod +x ${certScript}`);
 
   if (options['install-module']) {
-    chdir(`${distDir}/extras/basic_example`);
+    chdir(`${distDir}/apps/current_app`);
     execSync('npm install' + npmInstallOption);
   }
 }
@@ -627,7 +660,7 @@ getTargets()
   .then(cleanIfRepack)
   .then(processTargets)
   .then(packScripts)
-  .then(packSamples)
+  .then(packApps)
   .then(archive)
   .then(() => {
     console.log('\x1b[32mWork finished in directory:', distDir, '\x1b[0m');
